rapidfeedback tickets 1346 1342 1341 1340 1339 and some sanctions fixes

+++ extensions/content/rapidfeedback/classes/commands/AdminAction.class.php
@@ -32,14 +32,22 @@ class AdminAction extends \AbstractCommand implements \IAjaxCommand {
 					// copy
 					$copy = \steam_factory::create_copy($GLOBALS["STEAM"]->get_id(), $element);
 					$copy->move($rapidfeedback);
-					$copy->set_attribute("RAPIDFEEDBACK_PARTICIPANTS", array());
 					$copy->set_attribute("RAPIDFEEDBACK_STATE", 0);
-					$copy->set_attribute("RAPIDFEEDBACK_RESULTS", 0);
 					$copy->set_attribute("RAPIDFEEDBACK_STARTTYPE", 0);
+					
 					$resultContainer = \steam_factory::get_object_by_name($GLOBALS["STEAM"]->get_id(), $copy->get_path() . "/results");
-					$results = $resultContainer->get_inventory();
-					foreach ($results as $result) {
-						$result->delete();
+					$resultContainer->set_attribute("RAPIDFEEDBACK_RESULTS", 0);
+					$resultContainer->set_attribute("RAPIDFEEDBACK_PARTICIPANTS", array());
+					// clean resultcontainer and set sanctions
+					if ($resultContainer instanceof \steam_container) {
+						$results = $resultContainer->get_inventory();
+						foreach ($results as $result) {
+							$result->delete();
+						}
+						$groups = $rapidfeedback->get_attribute("RAPIDFEEDBACK_GROUP");
+						foreach ($groups as $group) {
+							$resultContainer->set_sanction($group, SANCTION_READ | SANCTION_WRITE | SANCTION_INSERT);
+						}
 					}
 					break;
 				case 4:
